chore: modularize for adding replay
